Resolve circular dependency in BCP
