+++ packages/iov-bcp/types/atomicswaptypes.d.ts
@@ -1,7 +1,58 @@
 import { As } from "type-tagger";
 import { Stream } from "xstream";
 import { BlockchainConnection } from "./connection";
-import { Address, Amount, SwapId, SwapTimeout } from "./transactions";
+import { Address, Amount, LightTransaction, SwapId, SwapTimeout } from "./transactions";
+export declare type Preimage = Uint8Array & As<"preimage">;
+export declare type Hash = Uint8Array & As<"hash">;
+/** A swap offer or a counter offer */
+export interface SwapOfferTransaction extends LightTransaction {
+  readonly kind: "bcp/swap_offer";
+  /**
+   * The ID of the swap to aid coordination between the two parties.
+   *
+   * If required, the data should be generated randomly by the client to avoid
+   * collisions.
+   *
+   * The type of this may be extended with additional properties depending on
+   * the requirements of the individual chain.
+   */
+  readonly swapId?: SwapId;
+  readonly amounts: readonly Amount[];
+  readonly recipient: Address;
+  /**
+   * The first point in time at which the offer is expired.
+   *
+   * Can be represented as a block height or UNIX timestamp.
+   */
+  readonly timeout: SwapTimeout;
+  /**
+   * Locally calculated hash of the preimage.
+   *
+   * This is a SHA256 hash until we have a way to specifiy the hashing algorithm.
+   */
+  readonly hash: Hash;
+  readonly memo?: string;
+}
+export interface SwapClaimTransaction extends LightTransaction {
+  readonly kind: "bcp/swap_claim";
+  readonly preimage: Preimage;
+  readonly swapId: SwapId;
+}
+export interface SwapAbortTransaction extends LightTransaction {
+  readonly kind: "bcp/swap_abort";
+  readonly swapId: SwapId;
+}
+export declare type SwapTransaction = SwapOfferTransaction | SwapClaimTransaction | SwapAbortTransaction;
+export declare function isSwapOfferTransaction(
+  transaction: LightTransaction,
+): transaction is SwapOfferTransaction;
+export declare function isSwapClaimTransaction(
+  transaction: LightTransaction,
+): transaction is SwapClaimTransaction;
+export declare function isSwapAbortTransaction(
+  transaction: LightTransaction,
+): transaction is SwapAbortTransaction;
+export declare function isSwapTransaction(transaction: LightTransaction): transaction is SwapTransaction;
 export declare enum SwapProcessState {
   Open = "open",
   Claimed = "claimed",
@@ -10,8 +61,6 @@ export declare enum SwapProcessState {
 export declare function isSwapProcessStateOpen(state: SwapProcessState): state is SwapProcessState.Open;
 export declare function isSwapProcessStateClaimed(state: SwapProcessState): state is SwapProcessState.Claimed;
 export declare function isSwapProcessStateAborted(state: SwapProcessState): state is SwapProcessState.Aborted;
-export declare type Preimage = Uint8Array & As<"preimage">;
-export declare type Hash = Uint8Array & As<"hash">;
 export interface SwapData {
   readonly id: SwapId;
   readonly sender: Address;
