+++ packages/iov-bcp/types/transactions.d.ts
@@ -1,5 +1,4 @@
 import { As } from "type-tagger";
-import { Hash, Preimage } from "./atomicswaptypes";
 export declare enum Algorithm {
   Ed25519 = "ed25519",
   Secp256k1 = "secp256k1",
@@ -199,53 +198,4 @@ export interface TimestampTimeout {
 }
 export declare function isTimestampTimeout(timeout: SwapTimeout): timeout is TimestampTimeout;
 export declare function createTimestampTimeout(secondsFromNow: number): TimestampTimeout;
-/** A swap offer or a counter offer */
-export interface SwapOfferTransaction extends LightTransaction {
-  readonly kind: "bcp/swap_offer";
-  /**
-   * The ID of the swap to aid coordination between the two parties.
-   *
-   * If required, the data should be generated randomly by the client to avoid
-   * collisions.
-   *
-   * The type of this may be extended with additional properties depending on
-   * the requirements of the individual chain.
-   */
-  readonly swapId?: SwapId;
-  readonly amounts: readonly Amount[];
-  readonly recipient: Address;
-  /**
-   * The first point in time at which the offer is expired.
-   *
-   * Can be represented as a block height or UNIX timestamp.
-   */
-  readonly timeout: SwapTimeout;
-  /**
-   * Locally calculated hash of the preimage.
-   *
-   * This is a SHA256 hash until we have a way to specifiy the hashing algorithm.
-   */
-  readonly hash: Hash;
-  readonly memo?: string;
-}
-export interface SwapClaimTransaction extends LightTransaction {
-  readonly kind: "bcp/swap_claim";
-  readonly preimage: Preimage;
-  readonly swapId: SwapId;
-}
-export interface SwapAbortTransaction extends LightTransaction {
-  readonly kind: "bcp/swap_abort";
-  readonly swapId: SwapId;
-}
-export declare type SwapTransaction = SwapOfferTransaction | SwapClaimTransaction | SwapAbortTransaction;
 export declare function isSendTransaction(transaction: LightTransaction): transaction is SendTransaction;
-export declare function isSwapOfferTransaction(
-  transaction: LightTransaction,
-): transaction is SwapOfferTransaction;
-export declare function isSwapClaimTransaction(
-  transaction: LightTransaction,
-): transaction is SwapClaimTransaction;
-export declare function isSwapAbortTransaction(
-  transaction: LightTransaction,
-): transaction is SwapAbortTransaction;
-export declare function isSwapTransaction(transaction: LightTransaction): transaction is SwapTransaction;
